requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
transformers==4.11.3
torch==1.10.0
numpy==1.21.2
pandas==1.3.3
requests==2.26.0